Create sub-window to show routing
Add a button to the main panel that opens a child window
that shows the amount of signal passing between each operator

Use color coding to the signal direction is more obvious.

The intentisty of the lines shows how much of that signal is
being used (so line not visible when 0)

Python/Tkinter/Canvas does not support opacity so a "trick"
is used. A "stipple" can be specified when the line is drawn to
set the density of the color. But to do this requires 256 levels
of stipple pattern and a "trick" that when stipple="" is used
an external XBM file can be given

So to make 256 different density stipple patterns start with
knobman and get it to make a 256 rectangle animated png where
each 32x32 rectangle is a different grey. Then use gimp to
convert this to 1 bit with Flloyd-Steinberg to create the stipple
patterns. Then write a small Python script that uses Pillow in kind
of the same way as the animation engine to crop the 32x32 rectangles
out of the large image and write each as XBM files.

Finally those XBM can be used at line drawing time to determine the
stipple/intensity level.

(Phew!)